requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pandas-ta>=0.4.67b0
yfinance>=0.2.50
dhanhq>=2.0.1
kiteconnect>=5.0.1
pyyaml>=6.0.2
python-dotenv>=1.0.1
streamlit>=1.40.0
plotly>=5.24.1
matplotlib>=3.9.0
python-telegram-bot>=21.7
apscheduler>=3.10.4
requests>=2.32.3
websocket-client>=1.8.0
pytz>=2024.2
pytest>=8.3.4
pyotp>=2.9.0
